Move HelpFormatter to core to keep systems off cli
Systems and v2 parsers were importing help formatting from cli,
inverting package layers. Keep a thin cli.help_format re-export for
compatibility.
