+++ src/kaist_cli/cli/parser.py
@@ -4,7 +4,7 @@ import argparse
 
 from ..core.distribution import discover_distribution_info
 from ..core.system_registry import default_registry
-from .help_format import HelpFormatter, dedent as _dedent
+from ..core.help_format import HelpFormatter, dedent as _dedent
 
 
 def build_parser() -> argparse.ArgumentParser:
